REFACTOR: move function to element
