+++ model/src/main/kotlin/vp/togedo/model/documents/personalSchedule/PersonalSchedule.kt
@@ -103,7 +103,7 @@ data class PersonalSchedule(
      * @throws PersonalScheduleEndTimeBeforeStartTimeException 종료 시간이 시작 시간보다 앞에 있음
      */
     fun addFixedPersonalScheduleElement(personalScheduleElement: PersonalScheduleElement): PersonalSchedule {
-        isValidTimeForFixedSchedule(personalScheduleElement)
+        personalScheduleElement.isValidTimeForFixedSchedule()
 
         val sortedIndex = getSortedIndex(
             personalScheduleElement = personalScheduleElement,
@@ -124,7 +124,7 @@ data class PersonalSchedule(
      */
     fun addFlexiblePersonalScheduleElement(personalScheduleElement: PersonalScheduleElement): PersonalSchedule {
 
-        isValidTimeForFlexibleSchedule(personalScheduleElement)
+        personalScheduleElement.isValidTimeForFlexibleSchedule()
 
         val sortedIndex = getSortedIndex(
             personalScheduleElement = personalScheduleElement,
@@ -168,58 +168,6 @@ data class PersonalSchedule(
                 throw ConflictPersonalScheduleException()
 
         return sortedIndex
-    }
-
-    /**
-     * 유동 스케줄의 시간이 유효한지 확인
-     * @param personalScheduleElement 확인할 요소
-     * @return true
-     * @throws PersonalScheduleTimeIsNotRangeException 유효한 시간 범위가 아님
-     * @throws PersonalScheduleEndTimeBeforeStartTimeException 종료 시간이 시작 시간보다 앞에 있음
-     */
-    fun isValidTimeForFlexibleSchedule(personalScheduleElement: PersonalScheduleElement): Boolean{
-        return personalScheduleElement.isStartTimeBefore() &&
-                //00(년)_01(월)_01(일)_00(시)_00(분) ~ 99(년)_12(월)_31(일)_23(시)_59(분)
-                isTimeRange(
-                    personalScheduleElement = personalScheduleElement,
-                    startTimeRange = "0001010000",
-                    endTimeRange = "9912312359",)
-    }
-
-    /**
-     * 고정 스케줄의 시간이 유효한지 확인
-     * @param personalScheduleElement 확인할 요소
-     * @return true
-     * @throws PersonalScheduleTimeIsNotRangeException 유효한 시간 범위가 아님
-     * @throws PersonalScheduleEndTimeBeforeStartTimeException 종료 시간이 시작 시간보다 앞에 있음
-     */
-    fun isValidTimeForFixedSchedule(personalScheduleElement: PersonalScheduleElement): Boolean{
-        return personalScheduleElement.isStartTimeBefore() &&
-                //1(요일)_00(시)_00(분) ~ 7(요일)_23(시)_59(분)
-                isTimeRange(personalScheduleElement = personalScheduleElement,
-                    startTimeRange = "10000",
-                    endTimeRange = "72359")
-    }
-
-    /**
-     * 해당 스케줄의 시간이 범위 내에 있는지 확인
-     * @param personalScheduleElement 확인할 요소
-     * @param startTimeRange 시작 범위
-     * @param endTimeRange 종료 범위
-     * @return true
-     * @throws PersonalScheduleTimeIsNotRangeException 유효한 시간 범위가 아님
-     */
-    private fun isTimeRange(
-        personalScheduleElement: PersonalScheduleElement,
-        startTimeRange: String,
-        endTimeRange: String): Boolean{
-        if(personalScheduleElement.startTime.length != startTimeRange.length ||
-            personalScheduleElement.endTime.length != endTimeRange.length ||
-            personalScheduleElement.startTime !in startTimeRange..endTimeRange ||
-            personalScheduleElement.endTime !in startTimeRange..endTimeRange){
-            throw PersonalScheduleTimeIsNotRangeException()
-        }
-        return true
     }
 
     /**
